requirements.txt
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
ecdsa==0.18.0
fastapi==0.105.0
greenlet==3.0.2
h11==0.14.0
httptools==0.6.1
idna==3.6
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
numpy==1.26.2
packaging==23.2
pandas==2.1.4
passlib==1.7.4
patsy==0.5.4
plotly==5.18.0
plotly-express==0.4.1
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.9
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
statsmodels==0.14.1
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2023.3
uvicorn==0.24.0.post1
# uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
